refactored merged_node -> MergedBlock -> MergedGraph
... and moved its implementation to Graph.hpp since its semantic behaviour is closer to a compile-time merged Graph than a Block.

* missing features to be addressed:
  * handling of tag forwarding
  * handling of sub-block settings
  * topology of internal graph (blocks+connections)
